requirements.txt
yt-dlp
openai-whisper
pysrt
torch
ffmpeg
